Environment tested with BCI headset, fixed direction bias problem, added Prompt Node for visual outputs. Started Launch file

+++ nav_goals/include/nav_goals/commandreader.h
@@ -17,7 +17,7 @@ class CommandReader
 public:
     CommandReader(ros::NodeHandle nh);
     void commandCallback(const std_msgs::Int64ConstPtr &message);
-    int getCounterWeight();
+    double getCounterWeight();
     int getDirection();
     int getOrientation();
     bool getEndFlag();
